chore: initial commit of prototype of new nav
