+++ package.json
@@ -59,6 +59,7 @@
     "@babel/runtime-corejs3": "7.21.0",
     "@electron/asar": "^3.2.3",
     "@fortawesome/fontawesome-svg-core": "^6.3.0",
+    "@fortawesome/free-brands-svg-icons": "^6.3.0",
     "@fortawesome/free-regular-svg-icons": "^6.3.0",
     "@fortawesome/free-solid-svg-icons": "^6.3.0",
     "@fortawesome/vue-fontawesome": "latest-2",
