+++ src/renderer/assets/scss/main.scss
@@ -3,14 +3,14 @@ html {
   user-select: none;
 }
 
-.page-enter-active,
-.page-leave-active {
-  transition: opacity 0.1s;
-}
-.page-enter,
-.page-leave-to {
-  opacity: 0;
-}
+// .page-enter-active,
+// .page-leave-active {
+//   transition: opacity 0.1s;
+// }
+// .page-enter,
+// .page-leave-to {
+//   opacity: 0;
+// }
 
 .pulse-danger {
   animation: pulseDanger 1s infinite;
