+++ src/renderer/pages/index.vue
@@ -58,16 +58,16 @@
           @change="resetColors()"
         />
       </v-col>
-      <v-col class="d-flex justify-end">
+      <!-- <v-col class="d-flex justify-end">
         <div class="mr-2">
           <shuffle-btn v-if="$getPrefs('meeting.enableMusicButton')" />
-        </div>
-        <template v-if="$getPrefs('media.enableMediaDisplayButton')">
-          <icon-btn variant="toggleScreen" class="mr-2" />
-          <icon-btn variant="present" :disabled="loading" class="mr-2" />
-        </template>
-        <icon-btn variant="settings" :disabled="loading" />
-      </v-col>
+        </div> -->
+        <!-- <template v-if="$getPrefs('media.enableMediaDisplayButton')"> -->
+          <!-- <icon-btn variant="toggleScreen" class="mr-2" /> -->
+          <!-- <icon-btn variant="present" :disabled="loading" class="mr-2" /> -->
+        <!-- </template> -->
+        <!-- <icon-btn variant="settings" :disabled="loading" /> -->
+      <!-- </v-col> -->
       <v-progress-linear
         v-if="currentProgress || totalProgress"
         fixed
@@ -613,6 +613,7 @@ export default defineComponent({
     color: #000 !important;
   }
 }
+
 .theme--dark {
   .secondary * {
     color: #fff !important;
